Modificaciones en la capa de servicios para crear un pasaje a partir de la existencia de un cliente y un paquete turístico, también en el código al aplicar el descuento.

+++ src/main/java/faz/api/svrp/dtos/PassageDto.java
@@ -4,7 +4,7 @@ public class PassageDto {
     private String date;
     private String ticketNumber;
     private String seatNumber;
-    private double totalPrice;
+    private float totalPrice;
     private int clientId;
     private int tourpackageId;
 
@@ -20,7 +20,7 @@ public class PassageDto {
         return seatNumber;
     }
 
-    public double getTotalPrice() {
+    public float getTotalPrice() {
         return totalPrice;
     }
